update opencode files and combine digital consent test
