+++ api/helpers/profiles/line.ts
@@ -35,14 +35,18 @@ const cleanupEmployee: CleanupStep = async (request, ctx) => {
     return
   }
 
-  // Pre-seed phase: find any employee left from a failed run and hard delete it
-  if (ctx.identifiers.email) {
+  // Pre-seed phase: find any employee left from a failed run and hard delete it.
+  // employee_id is generated with the EMPAPI prefix — unique per run and never
+  // used by manual testers, so it safely identifies automated test leftovers.
+  // email and line_id are not used: email changes every run, line_id is shared
+  // with manual test accounts and deleting by it would remove their data.
+  if (ctx.identifiers.employee_id) {
     try {
       const found = await findEmployeeByIdentifier(
         request,
         ctx.company,
-        ctx.identifiers.email,
-        'email'
+        ctx.identifiers.employee_id,
+        'employee_id'
       )
       if (found?.user_id) {
         try {
